Use new rebuildCache() function

+++ www/go/core/App.php
@@ -283,16 +283,19 @@ namespace go\core {
 				$this->rebuildCacheOnDestruct = $onDestruct;
 			}
 			
-			\GO::clearCache(); //legacy
-			
-			GO()->getCache()->flush(false);
-			db\Table::destroyInstances();
-	
-			$webclient = new \go\core\webclient\Extjs3();
-			$webclient->flushCache();
+			$lock = new util\Lock("rebuildCache");
+			if($lock->lock()) {
+				\GO::clearCache(); //legacy
 
-			\GO\Base\Observable::cacheListeners();
-			\go\core\event\Listeners::get()->init();
+				GO()->getCache()->flush(false);
+				db\Table::destroyInstances();
+
+				$webclient = new \go\core\webclient\Extjs3();
+				$webclient->flushCache();
+
+				\GO\Base\Observable::cacheListeners();
+				\go\core\event\Listeners::get()->init();
+			}
 		}
 		
 		public function __destruct() {
